+++ rust-spice/src/core/neat.rs
@@ -5,18 +5,22 @@ Improvement on the procedurally generated functions.
 
 The idiomatic Rust bindings to CSPICE can be very hard to generate in a procedural macro in some
 specific cases. You can find, in this module, fonctions wrapped from [`raw`] to better match
-an idiomatic usage. The improvements consists in functions taking:
+an idiomatic usage. The improvements consists in functions:
 
-+ a string as input in C requires to also send the size of the pointer to a char array. In Rust, you
++ taking a string as input in C requires to also send the size of the pointer to a char array. In Rust, you
   only send the string.
-+ taking input for array size and outputing size whereas a vector can be used
++ taking taking input for array size and outputing size whereas a vector can be used
++ which ouputs string that be allocated from default length sometimes
 */
 
 use crate::raw;
+use crate::MAX_LEN_OUT;
 
 /**
 This routine converts an input epoch represented in TDB seconds past the TDB epoch of J2000 to a
 character string formatted to the specifications of a user's format picture.
+
+See [`raw::dskp02`] for the raw interface.
 */
 pub fn timout<S>(et: f64, pictur: S) -> String
 where
@@ -28,6 +32,8 @@ where
 
 /**
 Fetch triangular plates from a type 2 DSK segment.
+
+See [`raw::dskp02`] for the raw interface.
 */
 pub fn dskp02(handle: i32, dladsc: raw::DLADSC) -> Vec<[i32; 3]> {
     let (_, np) = raw::dskz02(handle, dladsc);
@@ -36,6 +42,8 @@ pub fn dskp02(handle: i32, dladsc: raw::DLADSC) -> Vec<[i32; 3]> {
 
 /**
 Fetch vertices from a type 2 DSK segment.
+
+See [`raw::dskv02`] for the raw interface.
 */
 pub fn dskv02(handle: i32, dladsc: raw::DLADSC) -> Vec<[f64; 3]> {
     let (nv, _) = raw::dskz02(handle, dladsc);
@@ -43,15 +51,16 @@ pub fn dskv02(handle: i32, dladsc: raw::DLADSC) -> Vec<[f64; 3]> {
 }
 
 /**
-Return the current number of kernels that have been loaded via the KEEPER interface that are of
-a specified type.
+Fetch vertices from a type 2 DSK segment.
+
+See [`raw::kdata`] for the raw interface.
 */
 pub fn kdata<S: Into<String>>(which: i32, kind: S) -> (String, String, String, i32, bool) {
     raw::kdata(
         which,
         kind,
-        super::MAX_LEN_OUT as i32,
-        super::MAX_LEN_OUT as i32,
-        super::MAX_LEN_OUT as i32,
+        MAX_LEN_OUT as i32,
+        MAX_LEN_OUT as i32,
+        MAX_LEN_OUT as i32,
     )
 }